Fikset logikk slik at reduksjon begrunnelser ikke viser på vanlige perioder og litt refaktoering
