+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/vedtak/vedtaksperiode/VedtaksperiodeUtil.kt
@@ -3,6 +3,8 @@ package no.nav.familie.ba.sak.kjerne.vedtak.vedtaksperiode
 import no.nav.familie.ba.sak.common.FunksjonellFeil
 import no.nav.familie.ba.sak.common.TIDENES_MORGEN
 import no.nav.familie.ba.sak.common.erDagenFør
+import no.nav.familie.ba.sak.common.isSameOrAfter
+import no.nav.familie.ba.sak.common.isSameOrBefore
 import no.nav.familie.ba.sak.common.sisteDagIInneværendeMåned
 import no.nav.familie.ba.sak.common.toYearMonth
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingResultat
@@ -138,6 +140,69 @@ fun erFørsteVedtaksperiodePåFagsak(
     it.stønadFom.isBefore(
         periodeFom?.toYearMonth() ?: TIDENES_MORGEN.toYearMonth()
     )
+}
+
+fun identifiserReduksjonsperioder(
+    forrigeAndelerTilkjentYtelse: List<AndelTilkjentYtelse>,
+    andelerTilkjentYtelse: List<AndelTilkjentYtelse>,
+    vedtak: Vedtak,
+    opphørsperioder: List<VedtaksperiodeMedBegrunnelser>
+): List<VedtaksperiodeMedBegrunnelser> {
+    val forrigeSegmenter = forrigeAndelerTilkjentYtelse.lagVertikaleSegmenter().keys
+    val nåværendeSegmenter = andelerTilkjentYtelse.lagVertikaleSegmenter().keys
+    val segmenter = forrigeSegmenter.filterNot {
+        nåværendeSegmenter.any { seg -> it.fom == seg.fom && it.tom == seg.tom && it.value == seg.value }
+    }
+    val reduksjonsperioder = mutableListOf<VedtaksperiodeMedBegrunnelser>()
+    segmenter.filter { nåværendeSegmenter.any { seg -> seg.overlapper(it) } }
+        .forEach {
+            val overlappendePerioder =
+                nåværendeSegmenter.filter { nåSegment -> nåSegment.overlapper(it) && nåSegment.value < it.value }
+            overlappendePerioder.forEach { overlappendePeriode ->
+                reduksjonsperioder.add(
+                    VedtaksperiodeMedBegrunnelser(
+                        vedtak = vedtak,
+                        fom = if (it.fom > overlappendePeriode.fom) it.fom else overlappendePeriode.fom,
+                        tom = if (it.tom > overlappendePeriode.tom) overlappendePeriode.tom else it.tom,
+                        type = Vedtaksperiodetype.REDUKSJON,
+                    )
+                )
+            }
+        }
+
+    // opphørsperioder kan ikke være inkludert i reduksjonsperioder
+    return reduksjonsperioder.filterNot { reduksjonsperiode ->
+        opphørsperioder.any { it.fom == reduksjonsperiode.fom || it.tom == reduksjonsperiode.tom }
+    }
+}
+
+fun finnOgOppdaterOverlappendeUtbetalingsperiode(
+    utbetalingsperioder: List<VedtaksperiodeMedBegrunnelser>,
+    reduksjonsperioder: List<VedtaksperiodeMedBegrunnelser>
+): List<VedtaksperiodeMedBegrunnelser> {
+    val overlappendePerioder =
+        utbetalingsperioder.filter {
+            reduksjonsperioder.any { rd ->
+                rd.fom!!.isSameOrAfter(it.fom!!) && rd.tom!!.isSameOrBefore(
+                    it.tom!!
+                )
+            }
+        }
+    val oppdatertUtbetalingsperioder = mutableListOf<VedtaksperiodeMedBegrunnelser>()
+    utbetalingsperioder.forEach {
+        val overlappendePeriode = overlappendePerioder.firstOrNull { o -> it.fom == o.fom && it.tom == o.tom }
+        if (overlappendePeriode != null) {
+            oppdatertUtbetalingsperioder.addAll(
+                reduksjonsperioder.filter { r ->
+                    r.fom!! >= overlappendePeriode.fom &&
+                        r.tom!! <= overlappendePeriode.tom
+                }
+            )
+        } else {
+            oppdatertUtbetalingsperioder.add(it)
+        }
+    }
+    return oppdatertUtbetalingsperioder.sortedBy { it.fom }
 }
 
 fun hentGyldigeBegrunnelserForVedtaksperiode(
